Organize query test helpers.
