+++ src/components/__tests__/ContextView.ts
@@ -6,7 +6,7 @@ import store from '../../stores/app'
 import createTestApp, { cleanupTestApp } from '../../test-helpers/createRtlTestApp'
 import { findAllThoughtsByText } from '../../test-helpers/queries/findAllThoughtsByText'
 import { findSubthoughts } from '../../test-helpers/queries/findSubthoughts'
-import { findThoughtByText } from '../../test-helpers/queries/findThoughtByText'
+import findThoughtByText from '../../test-helpers/queries/findThoughtByText'
 import { getClosestByLabel } from '../../test-helpers/queries/getClosestByLabel'
 import { queryThoughtByText } from '../../test-helpers/queries/queryThoughtByText'
 import { setCursorFirstMatchActionCreator as setCursor } from '../../test-helpers/setCursorFirstMatch'
