+++ src/shortcuts/__tests__/toggleSort.ts
@@ -17,9 +17,9 @@ import createTestApp, { cleanupTestApp } from '../../test-helpers/createRtlTestA
 import { createTestStore } from '../../test-helpers/createTestStore'
 import { deleteThoughtAtFirstMatchActionCreator } from '../../test-helpers/deleteThoughtAtFirstMatch'
 import executeShortcut from '../../test-helpers/executeShortcut'
-import { findThoughtByText } from '../../test-helpers/queries'
-import { getDescendantsOfContext } from '../../test-helpers/queries/getDescendantsOfContext'
-import { getThoughtByContext } from '../../test-helpers/queries/getThoughtByContext'
+import findThoughtByText from '../../test-helpers/queries/findThoughtByText'
+import getDescendantsOfContext from '../../test-helpers/queries/getDescendantsOfContext'
+import getThoughtByContext from '../../test-helpers/queries/getThoughtByContext'
 import { setCursorFirstMatchActionCreator as setCursor } from '../../test-helpers/setCursorFirstMatch'
 import toggleSortShortcut from '../toggleSort'
