+++ src/test-helpers/queries/getBulletByContext.ts
@@ -4,10 +4,12 @@ import store from '../../stores/app'
 import hashPath from '../../util/hashPath'
 
 /** Get the bullet of a thought with the context. */
-export const getBulletByContext = (context: string[]): HTMLElement => {
+const getBulletByContext = (context: string[]): HTMLElement => {
   const path = contextToPath(store.getState(), context)
   const pathOfThought = hashPath(path)
   const bulletOfThought = screen.getByTestId('bullet-' + pathOfThought)
 
   return bulletOfThought
 }
+
+export default getBulletByContext
