+++ src/test-helpers/queries/getThoughtByContext.ts
@@ -4,10 +4,12 @@ import store from '../../stores/app'
 import hashPath from '../../util/hashPath'
 
 /** Get a thought with the context. */
-export const getThoughtByContext = (context: string[]): HTMLElement => {
+const getThoughtByContext = (context: string[]): HTMLElement => {
   const path = contextToPath(store.getState(), context)
   const pathOfThought = hashPath(path)
   const thought = screen.getByTestId('thought-' + pathOfThought)
 
   return thought
 }
+
+export default getThoughtByContext
