--- src/test-helpers/queries/index.ts
@@ -1,4 +1,0 @@
-/** All custom queries based on RTL. */
-
-export { findThoughtByText } from './findThoughtByText'
-export { findAllThoughtsByText } from './findAllThoughtsByText'
